Include the code changes from the May 14 stream
During the course of the stream, the demo was taken to the next level of
functionality with the inclusion of animations and a new sprite wrapper.

Instead of having an emote sprite parented to a parachute sprite, we now
have a special class for containing multiple sprites which is the parent
of both.

This allows us to apply animations to the outer parent container that
apply automatically to the children.

At the end of the stream, the implementation drops a random emote from
off the top of the screen, deploys the parachute and slowly wobbles as
it descends, then removes the parachute and the sway when touchdown
happens.