Move the utility code into a static Utils class
Due to a lack of desire for wanting loose functions floating around, the
utility code has been pushed into a class as static methods. The extras
that were added in for the vector code we didn't get around to adding
have been redacted pending the inclusion of that code at a future point.