Include some ability to play sounds
Include some simple sound effects for various aspects of the game. This
includes the sound of a parachute deploying, some landing sounds, an
occasional wilhelm scream as an emote enters for the drop, and a
whoopie sound for someone that scores.

The sounds play with a randomized playback speed in order to provide
some audible variety.

There are also currently a selection of landing sounds, though in
practice you would want to choose one appropriate to the landing
target.