Include a new sprite sheet for a target
Include a new sprite sheet (which currently contains only a single
sprite) for a target that will be where the incoming droppers are trying
to get to.

This also adds in a single target sprite using this sheet, positioned
randomly at the bottom of the screen with a slight gap to allow for an
emote to potentially land on the edges and hang off the side. This is
positioned slightly off the bottom of the screen as well, which looks
nicer and also hides the janky shading on the edge of the target.