Changing the collision bounding of the dropper
Previously, the collision volume used for the dropper were the full
bounds of the dropper itself, which means that the parachute would seem
to collide with the edges of the screen.

We now use the emote as the collision volume instead, since that's what
the dropper is doing.

This is not pixel perfect, particularly since the sway animation can
make things appear visually in a position their bounding rectangle does
not follow.

For the time being, lets assume that for a twitch stream game we don't
need to try to do appropriate rotation complete pixel perfect collision
bounding.

This also adds in collisions with the new target, allowing for scoring.
This is based off of how far from the center of the target the center of
the emote lands, with a perfect center being 100 points and a single
pixel overhang being roughly 0.44 points, give or take.