Include some death animation for looooosers
We now track whether a dropper is a winner or a loser based on where it
landed, and use that to keep winners around and get rid of losers.

Currently, losers are "ghosts" that are translucent, and after a delay
they slowly lower out of the screen while fading out.