Implement "The Highlander"
Implement the ability for the Target on the screen to know which
droppers are sitting on top of it at any given point. Whenever there are
two or more it performs the quickening and keeps only the one with the
highest score, telling the others that they are losers.

The result is that only the dropper with the highest score remains.