Perform some code cleanup
As we near the point where we're ready to integrate this code base with
the bot and the Twitch API, it behooves us to clean up the code and
make it somewhat less manky.

For this first step we're creating a new Entity class that will be the
base for everything that's displayed on the screen, and all of the
other classes inherit from it.