Rework the dropper to be re-usable
This is not fully tested yet, but the code for making a dropper and
filling it out as been moved inside of the constructor for the dropper
object and split into two methods; those which do the initial setup in
the constructor, and that which does initialization for the drop.

The idea is that we can remove a dropper from the DOM once it's dead,
and then re-use it later by just re-initializing it's internals.