Include an entity pool to reuse droppers
This implements the logic required to mark a dropper as being dead when
it has decided it should no longer be on the screen, add it to a pool
of previously created droppers to be reused, and remove itself from the
update loop.

To do this, the dropper listens for animations that are ending so that
it can tell when the animation that removes it from the screen has been
completed. This could also be extended to take other actions, such as
when the parachute is fully deployed.