Fix issues with reusing droppers
This resolves a small issue with reusing droppers in that the state of
the classes applied may not have been properly cleared (particularly for
droppers that had scores attached), which is visually disconcerting.

This resolves that issue, and also assigns to the overall dropper the
name associated with it as a single field, so that external code can
get at it if it wants to.

This is important for integration with Twitch, where you may want to
constrain droppers to one per user.