Separate the drop and cut actions
In order to allow for the bot or other external code in general to
control the drop and cut with more granularity, this simple change
splits this functionality out into two separate methods.

This could be enhanced somewhat by allowing the cut to happen only in
certain cases, or for drops to be allowed even if they're already
running, etc.