Add a randomized time delay to the cut operation
In order to make sure that it's not too easy for someone to bulleye when
they cut their chute, impose a random time delay between when the
operation is requested and when it actually triggers.

This includes an animation that will make it clearer to anyone watching
that a drop is in progress by having the emote visibly drop down
slightly when the cut happens.