Allow the render loop to control it's own destiny
The render loop has been modified so that it will only run while there
are active drops happening and for a brief period after the last drop
finishes, to allow for more drops.

When the loop isn't running, the target is hidden, and it comes back
when the loop restarts.