Include an endpoint to report results
This implements a single spot at which all results for a drop will be
reported so that external code that cares can be told about the results
of a drop, for example to report in a chat.

This is a blank stub since how you notify the back end relies on how you
interface this with your stream (if at all).

The contract here is that onTarget will be true if this dropped landed
or or got ejected from the target, and winner will be true if this
dropper is currently considered to be the winner.

As currently implemented, this gets triggered as "true, true" (on the
target and a winner) immediately when the dropper lands, but if there
is already a dropper on the target with a higher score, this will
immediately trigger again as "true, false" to indicate that you are not
the winner any longer.

This allows external code to be notified when someone gets knocked off
if they want to handle that situation.