Streamline the sample code a bit
In order to make the code easier to deploy to the bot from this stub,
shift the initialization code for the UI out into the web page, where
it can be easily elided as needed.

In addition, the initialization code is now in the top level of the
code file, where it will create the engine object to set everything
up.

This causes a single global (so that the UI can access it) although it
would be fairly easy to make that no be the case.

For the purposes of a simple twitch game, that seems like overkill given
the current situation.