Clarify (and fix?) the idle timer
This clarifies the exact situation in which the game should be able to
tell if it's idle or not, since at the end of the stream where this was
implemented we ended up with some frozen droppers still on the screen.

I'm not entirely sure why that occurred as it did, but this makes the
intention clearer, which for a demonstration game like this is worth
the change.