Fix an issue with restarting the render loop
There was a rather insidious bug that would stop the render loop
prematurely because the idle time is incremented by the delta time
between the current frame and the previous frame.

If the time of the last frame is not reset when the render loop
restarts, the first update will have a large deltaT that spans all of
the time since the loop started, in one update.

While technically true, that causes the idle time to immediately trip
again if it's been long enough.