Include a cut threshold with buzzer
This includes a new setting that specifies a point on the screen past
which it's not possible to cut any longer, which defaults to the bottom
of the screen.

If an attempt to cut is made past this point OR when a cut has already
occurred, then a buzz is played.