Make the sprite sheet dimensions be configurable
In order to make it easier for a user to replace the standard images
used with their own custom images, this simple modification sets up a
configuration item for the dimensions and sprites used in the sheets.

The actual specification of the images used comes from the CSS file,
but the code needs to know the dimensions of the sheet images and how
many sprites are available in it for randomly selecting sprites.