Include the ability to abdicate the target
Say that three times fast; this allows a winner that's currently on the
target to kick themselves off so that they can continue playing. In the
normal course of the game the only way off would be if another player
gets a higher score, or the game idles out.

There is a configuration option to determine if this is allowed or not,
which defaults to true.