Stop the cut buzzer on already dropped droppers
It was possible to attempt to cut a dropper that had already landed,
which results in the sound playing when it really should not, since
it's meant more for being a part of the overall drop mechanism.

This change stops the sound from playing if the dropper has already
landed.