Support testing with more than one dropper
This is currently fairly simplistic, but the idea is that since the
engine doesn't let someone drop if they're already dropping or on the
target, we need a way to allow for more than one dropper to exist.

For expedience in testing, we now add on a numeric suffix to all dropper
names that aren't explicitly provided, and there's a new button to
manually bump the number.

The naming of this button is sub-optimal.