Modify how the drop game reports winners
Previously every time any dropper landed on the target it would assume
it was a winner and transmit a message as such. Frequently this would
be followed immediately by a message saying that the dropper is now a
loser because someone on the target has a higher score.

This refactors things a bit so that the notification of a winner is only
triggered when a dropper actively gets a higher score (or is the only
dropper on the target); drops that land on the target with a lower
score will still generate a message, so that the back end can track
scores.

This is done by having the target track what it thinks the winning
dropper is and updating it accordingly, and deferring part of the
handling from the dropper to the target, which is the only thing that
knows about all of the droppers.