Fix issues with undersized twitch emotes
The twitch emote guideline specifies that the middle of the three
possible emote sizes is very specific; however, some global emotes such
as the ones for :) and similar aren't the correct size.

Previously this was an issue because the background image would tile in
one or both directions.

Now, we are sizing the image so that it's large enough to be contained
wholly within the container with no repeat.

In theory this should always be centered horizontally and at the top,
but it's hard to say for sure because despite the documentation for
background-position stating otherwise, setting it to anything seems to
have absolutely zero effect no matter what.

So, time will tell I guess. I suspect it will always work so long as
undersized images are taller than they are wide.